pyproject.toml
[project]
name = "homut-bot"
version = "0.1.0"
description = "Telegram bot for warehouse inventory management"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=20.0",
    "telegram>=0.0.1",
    "update>=0.0.1",
    "validators>=0.34.0",
]
